Add ability to add master configuration for owners
Organizations with multiple repositories shouldn't have to configure
Hound for each repo. It is likely the configuration per organization
stays the same, with minor overrides in individual repos.

This allows users to provide a master configuration for all of the
repositories under an organization. The flow would be to start with the
linter's default configuration, merge in Hound defaults, merge in
organization master configuration, then merge individual repo
configuration.

This might offer a solution for ripping out Hound defaults. We can point
existing repo configs to a master legacy config, then offer instructions
on how to migrate off of it.